fix: do not convert DN strings back and forth (Issue univention/ucs#2653)
The DN strings have been converted back and forth between "str" and "DN"
in a few lines. This does not hurt for a couple of entries, but takes
ages if we deal with 100k uniqueMember entries.

The code is optimized for a change of a small amount of uniqueMember
entries. The larger the changeset the longer it takes.