Issue univention/documentation/nubus-docs#87: Replace UDM HTTP REST API content with reference to Nubus customization guide
The content about the UDM HTTP REST API has seen a major rework.
It considers the deployments for Kubernetes and the Appliance.
The content is now part of the Nubus Customization Guide.

To avoid the maintenance of duplicate content, I replaced the content
with a link to the Nubus Customization Guide

Issue univention/documentation/nubus-docs#87