test(veyon): Add delay after join
After switching to a serial `rename_and_join`, errors
showed up. They were ignored beforehand as they occur during
the RDP group setup and in the background and are not critical for
the job to succeed. Adding a delay between the join and the Windows
group modifications seems to be sufficient.

Issue ucsschool#1363