feat(docs): Add local build make targets
The Makefile in the doc/manual directory provides the following
additional make targets:

* docker-gettext
* docker-html
* docker-latexpdf
* docker-linkcheck
* docker-livehtml
* docker-spelling
* docker-update-po

All those targets pull and run the Sphinx container image with the
Sphinx build environment and run the respective make command, exit
the container and remove it.

To use those targets, you need to be in the directory of the respective
document.

Issue univention/documentation/docs-projects#217