fix(docs): Add interactive and tty to docker run call
For local documentation builds the docker run commands need to run in
interactive mode with tty. Otherwise, Ctrl-c isn't passed on making it
hard to interrupt the docker-livehtml make rule.

Without --interactive --tty the user would need to stop the docker
container by hand. With the options, a simple Ctrl-c is enough and the
process stops.

Issue univention/documentation/docs-projects#217