fix: defer LDAP connection establishment
The code set up logging and the LDAP connection within the constructor.
This made it impossible to reuse the logger of the class within
univention-radius-ntlm-auth. The ldapConnection property is now
filled on demand and therefore the connection setup is delayed.

Additionally the code now checks if /etc/freeradius.secret is accessible
and logs a proper error message if this is not the case.

Bug #58132 / Issue univention/dev/ucs#2833